Delete initialize_node
