+++ src/dynamics/thermal/node.hpp
@@ -145,4 +145,12 @@ class Node {
   void AssertNodeParams(void);
 };
 
+/**
+ * @fn InitNode
+ * @brief Initialize Node object from csv file
+ * @param[in] node_str: str read from csv file
+ * @return Node
+ */
+Node InitNode(const std::vector<std::string>& node_str);
+
 #endif  // S2E_DYNAMICS_THERMAL_NODE_HPP_
